# openapi.yaml
openapi: "3.1.0"
info:
  title: korpus
  version: "2.0.0"

paths:
  /healthz:
    get:
      operationId: Healthz
      responses:
        "200":
          description: OK
        "503":
          description: Service Unavailable

  /api/status:
    get:
      operationId: GetStatus
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusResponse"

  /api/clusters:
    get:
      operationId: ListClusters
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /api/groups:
    get:
      operationId: ListGroups
      parameters:
        - name: cluster
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /api/namespaces:
    get:
      operationId: ListNamespaces
      parameters:
        - name: cluster
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  /api/kinds:
    get:
      operationId: ListKinds
      parameters:
        - name: cluster
          in: query
          schema:
            type: string
        - name: group
          in: query
          schema:
            type: string
        - name: namespace
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KindInfo"

  /api/resource:
    get:
      operationId: GetResource
      parameters:
        - name: cluster
          in: query
          required: true
          schema:
            type: string
        - name: group
          in: query
          required: true
          schema:
            type: string
        - name: kind
          in: query
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          schema:
            type: string
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: string
        "404":
          description: Not Found

  /api/diff:
    get:
      operationId: GetDiff
      parameters:
        - name: cluster
          in: query
          required: true
          schema:
            type: string
        - name: group
          in: query
          required: true
          schema:
            type: string
        - name: kind
          in: query
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          schema:
            type: string
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: from
          in: query
          required: true
          schema:
            type: string
        - name: to
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiffResult"
        "400":
          description: Bad Request
        "404":
          description: Not Found

  /api/snapshot:
    get:
      operationId: GetSnapshot
      parameters:
        - name: datetime
          in: query
          schema:
            type: string
            format: date-time
        - name: cluster
          in: query
          schema:
            type: string
        - name: group
          in: query
          schema:
            type: string
        - name: kind
          in: query
          schema:
            type: string
        - name: namespace
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
        - name: cel
          in: query
          description: CEL expression (only valid when datetime is omitted)
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotPage"
        "400":
          description: Bad Request (e.g. cel combined with datetime)

  /api/history:
    get:
      operationId: GetHistory
      parameters:
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          schema:
            type: string
            format: date-time
        - name: cluster
          in: query
          schema:
            type: string
        - name: group
          in: query
          schema:
            type: string
        - name: kind
          in: query
          schema:
            type: string
        - name: namespace
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
        - name: changeType
          in: query
          schema:
            type: string
            enum: [added, modified, deleted]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoryPage"

  /api/volatility:
    get:
      operationId: GetVolatility
      parameters:
        - name: cluster
          in: query
          schema:
            type: string
        - name: group
          in: query
          schema:
            type: string
        - name: kind
          in: query
          schema:
            type: string
        - name: namespace
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
        - name: commits
          in: query
          schema:
            type: integer
            default: 50
        - name: threshold
          in: query
          schema:
            type: number
            default: 0.0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VolatilityPage"

  /api/volatility/fields:
    get:
      operationId: GetVolatilityFields
      parameters:
        - name: cluster
          in: query
          schema:
            type: string
        - name: group
          in: query
          required: true
          schema:
            type: string
        - name: kind
          in: query
          required: true
          schema:
            type: string
        - name: namespace
          in: query
          schema:
            type: string
        - name: name
          in: query
          schema:
            type: string
        - name: commits
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FieldVolatilityEntry"
        "400":
          description: Bad Request

components:
  schemas:
    KindInfo:
      type: object
      required:
        - group
        - kind
      properties:
        group:
          type: string
        kind:
          type: string

    SnapshotResource:
      type: object
      required:
        - cluster
        - group
        - kind
        - name
        - namespace
      properties:
        cluster:
          type: string
        group:
          type: string
        kind:
          type: string
        name:
          type: string
        namespace:
          type: string
        labels:
          nullable: true
          type: object
          additionalProperties:
            type: string
        creationTimestamp:
          type: string

    SnapshotPage:
      type: object
      required:
        - items
        - total
        - offset
        - limit
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SnapshotResource"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        commitSha:
          nullable: true
          type: string
        commitTime:
          nullable: true
          type: string
          format: date-time

    ChangeEvent:
      type: object
      required:
        - timestamp
        - sha
        - cluster
        - group
        - kind
        - namespace
        - name
        - changeType
      properties:
        timestamp:
          type: string
          format: date-time
        sha:
          type: string
        cluster:
          type: string
        group:
          type: string
        kind:
          type: string
        namespace:
          type: string
        name:
          type: string
        changeType:
          type: string
          enum: [added, modified, deleted]

    HistoryPage:
      type: object
      required:
        - items
        - total
        - offset
        - limit
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ChangeEvent"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer

    DiffResult:
      type: object
      required:
        - before
        - after
      properties:
        before:
          type: string
        after:
          type: string

    VolatilityEntry:
      type: object
      required:
        - cluster
        - group
        - kind
        - namespace
        - name
        - count
        - total
        - ratio
      properties:
        cluster:
          type: string
        group:
          type: string
        kind:
          type: string
        namespace:
          type: string
        name:
          type: string
        count:
          type: integer
        total:
          type: integer
        ratio:
          type: number

    VolatilityPage:
      type: object
      required:
        - items
        - total
        - offset
        - limit
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/VolatilityEntry"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer

    FieldVolatilityEntry:
      type: object
      required:
        - field
        - count
        - total
        - ratio
      properties:
        field:
          type: string
        count:
          type: integer
        total:
          type: integer
        ratio:
          type: number

    ClusterStatus:
      type: object
      required:
        - name
        - resourceCount
        - lastPullErr
      properties:
        name:
          type: string
        lastPullAt:
          nullable: true
          type: string
          format: date-time
        lastPullErr:
          type: string
        resourceCount:
          type: integer

    StatusResponse:
      type: object
      required:
        - clusters
      properties:
        clusters:
          type: array
          items:
            $ref: "#/components/schemas/ClusterStatus"